refactor: removing prev ObservationCollection
