+++ Packages/com.zelcam4.ml-agents/Runtime/Autoring/AgentAuthoring.cs
@@ -9,8 +9,6 @@ namespace Zelcam4.MLAgents
         public string behaviourName = "a";
         public int maxStep;
         
-        [Header("Observations")]
-        public ObservationSourceType[] observationSetup;
         
         [Header("Actions")]
         [Tooltip("The number of continuous actions the agent can take.")]
@@ -62,17 +60,6 @@ namespace Zelcam4.MLAgents
                     Target = GetEntity(authoring.target, TransformUsageFlags.Dynamic),
                     GoundRender = GetEntity(authoring.groundRender, TransformUsageFlags.None),
                 });
-                
-                // Observation baking
-                DynamicBuffer<ObservationValue> observations = AddBuffer<ObservationValue>(entity);
-                observations.ResizeUninitialized(authoring.observationSetup.Length);
-                
-                DynamicBuffer<ObservationSource> sources = AddBuffer<ObservationSource>(entity);
-                sources.Capacity = authoring.observationSetup.Length;
-                foreach (var sourceType in authoring.observationSetup)
-                {
-                    sources.Add(new ObservationSource { SourceType = sourceType });
-                }
                 
                 // Policy baking
                 AddComponent(entity, new RemotePolicy());
