--- Packages/com.zelcam4.ml-agents/Runtime/Systems/GatherObservationSystem.cs.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: 83cd02ce26db486096498d447692795d
-timeCreated: 1753895906
